move info to utils, update qtabout style
