+++ napari/_qt/qt_about.py
@@ -9,7 +9,7 @@ from qtpy.QtWidgets import (
     QHBoxLayout,
 )
 
-from .._info import sys_info, citation_text
+from ..utils import sys_info, citation_text
 
 
 class QtAbout(QDialog):
@@ -38,6 +38,10 @@ class QtAbout(QDialog):
         self.layout.addLayout(self.info_layout)
 
         self.infoTextBox.setText(sys_info(as_html=True))
+        self.infoTextBox.setMinimumSize(
+            self.infoTextBox.document().size().width() + 19,
+            self.infoTextBox.document().size().height() + 10,
+        )
 
         self.layout.addWidget(QLabel('<b>citation information:</b>'))
         self.citationTextBox = QTextEdit(citation_text)
