+++ napari/utils/__init__.py
@@ -1,0 +1,1 @@
+from .info import sys_info, citation_text
